# Conflicts:
#	src/main/java/edu/wpi/cs3733d18/teamp/ui/home/MainController.java
#	src/main/java/edu/wpi/cs3733d18/teamp/ui/service/RecordScreenController.java
#	src/main/resources/FXML/home/HomeScreen.fxml

+++ src/main/java/edu/wpi/cs3733d18/teamp/ui/admin/MapBuilderEdgeFormController.java
@@ -14,10 +14,20 @@ import edu.wpi.cs3733d18.teamp.Pathfinding.Node;
 import edu.wpi.cs3733d18.teamp.ui.admin.MapBuilderController;
 import javafx.event.ActionEvent;
 import javafx.fxml.FXML;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
 import javafx.scene.control.*;
 import javafx.scene.layout.GridPane;
+import javafx.stage.Modality;
+import javafx.stage.Stage;
+
+import java.io.IOException;
 
 public class MapBuilderEdgeFormController {
+
+    DeletePopUpController deletePopUpController;
+
     DBSystem db = DBSystem.getInstance();
 
     Boolean editFlag = false;
@@ -86,39 +96,37 @@ public class MapBuilderEdgeFormController {
     public void submitEdgeFormButtonOp(ActionEvent e) {
         Boolean isActive = isActiveEdgeCheckBox.isSelected();
         Node startNode, endNode;
+
+        // Validate both fields have been filled in by user
         try {
             startNode = db.getOneNode(startNodeTxt.getText());
             endNode = db.getOneNode(endNodeTxt.getText());
         } catch (NodeNotFoundException nnfe) {
-            nnfe.printStackTrace();
-            edgeFormErrorLabel.setText("Incompatible input for fields which require numbers.");
+            edgeFormErrorLabel.setText("Please choose two valid nodes.");
             edgeFormErrorLabel.setVisible(true);
+            System.out.println(nnfe.getNodeID());
             return;
         }
         Edge edge = new Edge(startNode, endNode, isActive);
 
-        if(editFlag){
-            edge.setID(editedEdgeID);
-            try {
+        // Validate an edge does not already exist between these nodes TODO probably not necessary
+
+        // After validation, create the edge
+        try {
+            if (editFlag) {
+                edge.setID(editedEdgeID);
                 db.modifyEdge(edge);
+                editedEdgeID = null;
             }
-            catch(NodeNotFoundException nnfe){
-                nnfe.printStackTrace();
-            }
-            catch(EdgeNotFoundException enfe){
-                enfe.printStackTrace();
-            }
-            editedEdgeID = null;
-        } else {
-            try {
+            else {
                 db.createEdge(edge);
             }
-            catch(NodeNotFoundException nnfe){
-                nnfe.printStackTrace();
-            }
-            catch(EdgeNotFoundException enfe){
-                enfe.printStackTrace();
-            }
+        }
+        catch(NodeNotFoundException nnfe){
+            nnfe.printStackTrace();
+        }
+        catch(EdgeNotFoundException enfe){
+            enfe.printStackTrace();
         }
 
         mapBuilderController.updateMap();
@@ -166,20 +174,48 @@ public class MapBuilderEdgeFormController {
      * Deletes an edge
      */
     public void deleteButtonOp(){
-        try{
+
+        // Check if deleting this edge will orphan a node
+        try {
             db.willEdgeOrphanANode(editedEdgeID);
         } catch (OrphanNodeException o) {
-            //edgeFormErrorLabel.setText("Deleting Edge will orphan Node: " + o.getNodeID());
-            //edgeFormErrorLabel.setVisible(true);
+            edgeFormErrorLabel.setText("Deleting Edge will orphan Node: " + o.getNodeID());
+            edgeFormErrorLabel.setVisible(true);
             return;
         }
+
+        // Delete Node Popup if edge can be deleted
+        Stage stage;
+        Parent root;
+        FXMLLoader loader;
+
+        stage = new Stage();
+        loader = new FXMLLoader(getClass().getResource("/FXML/general/ConfirmationPopUp.fxml"));
+
         try {
-            db.deleteEdge(editedEdgeID);
-        } catch (NodeNotFoundException | EdgeNotFoundException ne) {
-            ne.printStackTrace();
+            root = loader.load();
+        } catch (IOException ie) {
+            ie.printStackTrace();
+            return;
         }
-        mapBuilderController.updateMap();
-        mapBuilderController.addOverlay();
+
+        deletePopUpController = loader.getController();
+        deletePopUpController.StartUp(this);
+        stage.setScene(new Scene(root, 600, 150));
+        stage.initModality(Modality.APPLICATION_MODAL);
+        stage.initOwner(deleteButton.getScene().getWindow());
+        stage.showAndWait();
+
+        // If user confirms, delete edge
+        if (deletePopUpController.getChoice()) {
+            try {
+                db.deleteEdge(editedEdgeID);
+            } catch (NodeNotFoundException | EdgeNotFoundException ne) {
+                ne.printStackTrace();
+            }
+            mapBuilderController.updateMap();
+            mapBuilderController.addOverlay();
+        }
     }
 
 }
